abstracted the scraper to be customizable
