+++ examples/settings.json
@@ -3,9 +3,9 @@
     "esIndex": "hermes",
     "esType": "webpages",
     "crawlDelay": 5,
-    "cancelDuration": 0,
+    "cancelDuration": 15,
     "cancelUrl": "",
-    "stopDuration": 0,
+    "stopDuration": 15,
     "stopUrl": "",
     "memStatsInterval": 30,
     "userAgent": "(Hermes github.com/jtaylor32/hermes)",
